Change all long to int and remove int?

+++ src/Imi.Project.Api.Core/Interfaces/Repository/IRepository.cs
@@ -10,14 +10,14 @@ namespace Imi.Project.Api.Core.Interfaces.Repository
 {
     public interface IRepository<T> where T : BaseEntity
     {
-        Task<T> GetByIdAsync(long id);
-        Task<T> GetByIdAsync(long id, string[] includes);
+        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(int id, string[] includes);
         IQueryable<T> GetAllAsync();
         Task<IEnumerable<T>> ListAllAsync();
         Task<T> AddAsync(T entity);
         Task<IEnumerable<T>> AddRangeAsync(IEnumerable<T> entity);
         Task<T> UpdateAsync(T entity);
         Task<T> DeleteAsync(T entity);
-        Task<T> DeleteAsync(long id);
+        Task<T> DeleteAsync(int id);
     }
 }
