+++ src/Imi.Project.Api.Core/Interfaces/Repository/IWatchlistRepository.cs
@@ -6,8 +6,8 @@ namespace Imi.Project.Api.Core.Interfaces.Repository
 {
     public interface IWatchlistRepository : IRepository<Watchlist>
     {
-        Task<Watchlist> GetByUserIdAndMovieIdAsync(string userId, long movieId);
-        Task<Watchlist> AddByUserIdAndMovieIdAsync(string userId, long movieId);
+        Task<Watchlist> GetByUserIdAndMovieIdAsync(string userId, int movieId);
+        Task<Watchlist> AddByUserIdAndMovieIdAsync(string userId, int movieId);
         Task<IEnumerable<Watchlist>> GetWatchlistsByUserIdAsync(string userId);
     }
 }
