+++ src/Imi.Project.Api.Core/Interfaces/Service/IFavoriteService.cs
@@ -6,8 +6,8 @@ namespace Imi.Project.Api.Core.Interfaces.Service
 {
     public interface IFavoriteService
     {
-        Task<FavoriteResponseDto> GetByIdAsync(long id);
-        Task<FavoriteResponseDto> GetByUserIdAndMovieId(string userId, long movieId);
+        Task<FavoriteResponseDto> GetByIdAsync(int id);
+        Task<FavoriteResponseDto> GetByUserIdAndMovieId(string userId, int movieId);
         Task<FavoriteResponseDto> AddFavoriteAsync(FavoriteRequestDto favoriteRequestDto);
         Task DeleteFavoriteAsync(FavoriteRequestDto favoriteRequestDto);
         Task<IEnumerable<FavoriteResponseDto>> GetFavoritesByUserId(string userId);
