+++ src/Imi.Project.Api.Core/Interfaces/Service/IWatchlistService.cs
@@ -6,8 +6,8 @@ namespace Imi.Project.Api.Core.Interfaces.Service
 {
     public interface IWatchlistService
     {
-        Task<WatchlistResponseDto> GetByIdAsync(long id);
-        Task<WatchlistResponseDto> GetByUserIdAndMovieId(string userId, long movieId);
+        Task<WatchlistResponseDto> GetByIdAsync(int id);
+        Task<WatchlistResponseDto> GetByUserIdAndMovieId(string userId, int movieId);
         Task<WatchlistResponseDto> AddWatchlistAsync(WatchlistRequestDto watchlistRequestDto);
         Task DeleteWatchlistAsync(WatchlistRequestDto watchlistRequestDto);
         Task<IEnumerable<WatchlistResponseDto>> GetWatchlistsByUserId(string userId);
