+++ src/Imi.Project.Api.Infrastructure/Repositories/EfRepository.cs
@@ -20,12 +20,12 @@ namespace Imi.Project.Api.Infrastructure.Repositories
             _dbContext = dbContext;
         }
 
-        public virtual async Task<T> GetByIdAsync(long id)
+        public virtual async Task<T> GetByIdAsync(int id)
         {
             return await _dbContext.Set<T>().FindAsync(id);
         }
 
-        public virtual async Task<T> GetByIdAsync(long id, string[] includes)
+        public virtual async Task<T> GetByIdAsync(int id, string[] includes)
         {
             var query = _dbContext.Set<T>().AsQueryable();
 
@@ -74,7 +74,7 @@ namespace Imi.Project.Api.Infrastructure.Repositories
             return entity;
         }
 
-        public async Task<T> DeleteAsync(long id)
+        public async Task<T> DeleteAsync(int id)
         {
             var entity = await GetByIdAsync(id);
             await DeleteAsync(entity);
