+++ src/Imi.Project.Wpf.Core/Entities/MovieActor.cs
@@ -4,10 +4,10 @@ namespace Imi.Project.Wpf.Core.Entities
 {
     public class MovieActor : BaseEntity
     {
-        public long MovieId { get; set; }
+        public int MovieId { get; set; }
         public Movie Movie { get; set; }
 
-        public long ActorId { get; set; }
+        public int ActorId { get; set; }
         public Actor Actor { get; set; }
 
     }
